+++ generate-server-info.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 BRANCH_NAME=$(git branch | awk '{ print $2 }')
 COMMIT_ID=$(git rev-parse HEAD)
-echo "{ \"version\":\"${COMMIT_ID}\", \"branch\":\"${BRANCH_NAME}\", \"name\":\"flask-example-app\"}" > server_info.json
+echo "{ \"version\":\"${COMMIT_ID}\", \"branch\":\"${BRANCH_NAME}\", \"name\":\"flask-app-example\"}" > server_info.json
 echo "Wrote server_info.json"