+++ tests/test_root.py
@@ -1,12 +1,12 @@
 def test_index(client):
     response = client.get("/")
     data = response.get_json()
-    assert "flask-example-app" == data["name"]
+    assert "flask-app-example" == data["name"]
     assert "It works on my machine!" == data["message"]
 
 
 def test_server_info(client):
     response = client.get("/server_info")
     data = response.get_json()
-    assert "flask-example-app" == data["name"]
+    assert "flask-app-example" == data["name"]
     assert data["version"] is not None